+ hosts: add rpis and jumpgate method to reach
use a ProxyCommand to reach the Raspberry Pis via a jumpgate,
in this case the proxy command basically runs ssh to the jump
host and connects the stdin/stdout of the client (from the
perspective of the localmost initiator) to the far-side remote
ssh host and port (i.e. on the raspberry pi) and then we ssh
to those connected tunnels.  the result does basically a nested
ssh without the jumphost needing to make any connection itself
(similar to how we do backups via ssh -L which backgrounds
a port forward, and then connect through local end, but we
use stdout instead and all done in one command, everything
exits when the connection is done, instead of leaving around
a backup process).  we add these proxy commands as variables
for a group "jumpees" and then add our rpis to the group