+ rc: add local virt hosts so they get hostvars
we will be adding an ansible role that instantiates virtual machines,
and takes a host argument.  those hosts need to be in hostvars[] to pick
up variables from host_vars/, so we add them to inventory.  in future
these will be targets for further plays anyways, so might as well add
them now.