+ ark: new directory with superseded/onhold stuff
create an archive directory with content that we decided against using
and/or moved to other repositories, but we want the files visible in an
archive rather than having to search back through git history to find
them because we anticipate possibly some use or needing to reference it
later

first customer is the srcpush, to push all marked git repos to public
repositories.  we determined that doing this in ansible was too slow,
and we're just going to do it in shell instead.  however we might visit
this again later once #12086 'forks' has been implemented to allow local
parallelism, otherwise this is just too slow