+ initssh: add guard condition for initssh_keyid
rather than trying with files globbed from a name which won't expand if
the variable is defined, but still does a stat(), e.g.:

	files/{{initssh_keyid}}-id_rsa.pub

will try

	file/-id_rsa.pub

if initssh_keyid isn't set.  So we add a guard condition on that string
having a value.

(note, this might not be true, because the script doesn't return an error
when initssh_keyid isn't set, so this leads me to think it's never
actually doing the stat() but rather just bombing out when it cannot
expand the variable.  that said it's still safer/correct to put the
check in there ourselves)