- rc: switch to homedir-relative minimal config
instead of assuming checkout in /etc/ansible, the model we've come to
use is having the .ansible.cfg be in the homedir of the user we're
operating as, ie the homedir *is* an ansible checkout.  this works
better than a system-wide because it allows for multiple ansible
checkouts and configuration files in different locations for different
tasks/purposes on the same machine.

we also get rid of some optimizations like caching facts, by pulling in
the more minimal config file developed on some currently-worked ansible
projects (out of this tree).  there have been various issues with fact
caching, persistent ssh connections, etc and we've come to value the
correctness / determinism of disabling these optimizations, more than
the speed they provide.  this will also let us develop off a common base
config wrt our other projects.