^ hosts: remove inventory now that we use hostnames
we're using dns and applying roles with ansrole, which takes hostnames.
so, we don't need an inventory.  as well, our ansrole wrapper is setting
inventory to /dev/null, so this file isn't needed.  we'll use genders
and/or some other inventory script to generate inventory if we need
that later...