/ initpgp: gpg: only kill our own stray gpg-agent
gpg leaves a leftover agent we don't want on the system, and there's no
way to tell it not to do that, so we clean it up.

note that we're relying on the user being able to reload the agent with
his keys the next time it asks, this is normally fine except if we leave
some kind of long running job somewhere which requires the agent and
then re-run dotfiles setup and it kills our agents.  this seems very
unlikely, would take special arrangements if this was happening.