/ files: keys: make key correspond to whats passed from cli
from cli we just pass initpgp_keyid=scott@smemsh.net and this is not
looked up with keyid, it just tries for files/scott@smemsh.net.asc, and
that key was supposed to be the same as this with a -keyid suffix but
was not.

so make it a symlink instead of a copy.  should work in different
contexts where it's used such as the copy module, i.e. it won't copy a
symlink but its contents.  we'll see if it works