- files: keys: check in 2019 version of public key + sigs
this is the 2019 smemsh key, but it's already expired.  reason to add
this now is we're about to add the 2020 keys, which appear to be missing
earlier signatures, so want this to be checked in for posterity, in case
we need to go back for some reason.  don't think expired signatures are
needed anymore, but want to be complete.