- files: keys: check in 2020 version of gpg pubkey + sigs
these apparently are missing earlier signatures, but they don't appear
to be present on my keyring anymore.  there was some event where the
keyring had to be recovered from backup and it was probably exported
incorrectly, or something like that.  the key and signatures on it are
valid from today on, and this is only used for new installations, so
don't think it matters that the old expired keys are no longer present