- smemsh: meta: move role deps to tasks so can run pkg checks beforehand
we need to make sure 'acl' package is installed prior to running roles
that use it, and we don't have control at the playbook-level so we
cannot run a pre_tasks at this point.  the answer is to bring these in
via include_role and friends

used import_role when we didn't need to apply a 'when' clause, which has
different behavior of applying to all the tasks if 'import' variant is
used vs 'include' which applies the 'when' to the include itself.