- sysadmin: make the group if dne, else bail if gids don't match
this was never actually making the group (run condition was always
'false'), not exactly sure why.  never noticed since we always have
something before it creating the group.

since this block never ran, this was causing the variable
'sysadmin_name' to never to get referenced, so it never got lazy-loaded,
and downstream roles weren't able to access the 'sysadmin_name'
variable.  strangely, this didn't seem to happen when the role was run
via 'meta' rather than run by 'import_role'.

guessing this was set 'false' because it screwed up a system where the
gid was already set to something else, and the gid got changed.
unfortunately this is pre-history so the actual reason is lost.  it
could also just be a mistake but that's unlikely (why would there be a
random when: false condition?)