/ smemsh: make sysadmin role vars public for use in subsequent tasks
https://github.com/ansible/ansible/issues/63443 describes why we need to
do this.  it's a consequence of changing from meta: to putting them in
the play at runtime.  we could also have used import_role, which is
basically what 'role:' in meta/main.yml are, but that seemed funny since
it really just attaches the 'when' clause to every task in the play, and
get imported at parse-time.  if we're doing it at run-time then let's do
it at run-time...