^ initpgp: revert gpg warning silence so it works with more versions
we had added a commit that suppressed a warning, but in return that
caused it to not work on earlier versions of gpg.  the way to work with
2.0+ is to just let it emit the warning, it actually works anyways, and
has the right exit code, it just has the warning, which we can ignore
for now.  later, we might also consider using just

    --import --with-colons --import-options show-only

as described in
https://stackoverflow.com/questions/22136029/how-to-display-gpg-key-details-without-importing-it

which shouldn't list the extraneous signatures either