/ gitmodules: fix breakage by changing back to canonical urls
if we want to develop without going through github, we should not change
the canonical urls but rather the remotes used by the checkouts.
otherwise various issues are encountered.  so the initial bootstrap is
done using the canonical upstream url on github, then after checkout we
have to manually go into the checkout and change the remote for each one
to refer to the local filesystem.