- venv: makefile: peg the python version, as we depend on it
we are starting development after several stalled years and we have
issues restarting where we left off on the newer python.  so we'll work
in that upgrade gradually as we upgrade ansible itself.  first we have
to get everything working as it was at the point where we left off so
let's peg to the python major/minor version (still in /usr/local)