+ git: reltag: add our tool to update repository from latest tag
this is from https://github.com/smemsh/git-reltag which we use to keep
checkouts up to date with latest tag.  and would include this repo
itself.  rather than rely on the tool being installed already, or used
by sysadmin user, include submodule with the reltag repo and a symlink
(so installx/installrc do not have to be present)