+ tf: add lab environment implemented in terraform
start a single flat hcl file with the essentials of our lab env either
imported or specified.

we also include an external data source script to access a full run of
gethostent(3) records, which strangely is not available by default in
terraform so we have to make a data source.

this is more or less complete and working at the start and forms a good
base to start individual checkins and splitting up files, getting the
ansible working (it will be calling terraform)