= tf: main: add comment separator lines for logical sections
we will eventually split this into the "standard" terraform file-set and
these logical separations will form the basis of content divisions that
will end up in separate files.  it also helps while the file is a
monolith because it's growing in size and it's easier to read with the
separators.