+ tfcloudinit: python script to run ansible cloudinit role for tf data
this script runs the ansible 'cloudinit' role and captures the output
back as network-config and user-data variables in a terraform data
source. takes a single parameter which is the "target" node to render
the config for.

also includes a bit of terraform glue, to support the data importation
in terraform