- cloudinit: write_files for hosts and resolv.conf, not bootcmd
bootcmd is actually run every boot, which is not what we want.  it would
necessitate removing or disabling cloud-init, but we actually don't want
that.  write_files is run once per instance, not once, so we have to be
careful if iid changes and we haven't changed the entire cloud-config
(composite md5sum of which forms iid) but pretty sure that will be done
when we do the second boot (ie deploy boot, from image snapshot)