/ tfcloudinit: check for error ourselves so terraform can pretty-print
we were not sending back error correctly (it would exit 0 even on
error), this was fixed in upstream devskel

furthermore, terraform seems to only report it if it came from stderr,
which, despite several attempts, could not get working correctly with
check_output.  that just wraps run() so we just do it ourselves instead
and seems to work fine now.

so, this means we don't need to change code and trace things when
cloudinit rendering fails, but instead can easily just see it in the
output.  makes debugging faster...