^ rc: gather_subset: comment this out as it is no longer an rc var
this was marked deprecated and removed as a global config, now we're
supposed to set it at play or task level using "module options." we need
to look further into this and set it where appropriate, but the original
use case was mainly because fact gathering was actually causing
spontaneous reboot of our small office dns server embedded devices.

overall, fact gather should probably be explicit and we should say
exactly what facts we need, but for now we'll just go back to default
gathering behavior (since this is no longer recognized anyways) and
leave this to a later project.

there's all kinds of issues with fact gathering and scalability with
ansible... see the issue linked in the comments, for example