= cloudinit: todo: add note to try different var init sequences
we have an issue where we cannot see variables in inventory until a play
is run against a host.  thought this was just a known thing with ansible
and it had to be worked around by manually invoking add_host module, but
while working on something unrelated, discovered this setting.  adding a
reminder to check it out later.