- auth: set_password: use password_hash filter, not mkpasswd binary
as long as passlib python package is installed, this filter is
available, and works (tested and it worked fine).  there are some
caveats though, passlib has not been updated in 5 years, ansible devs
want to use hashlib maybe, but it doesn't cover all the algorithms.  see
ansible issue 82652 for details.