^ skelfiles: don't remove ~/.profile, it won't be there usually
no need to remove homedir files because that was already handled
out-of-band during a different stage of provisioning, at least in the
case of config user, which is the first consumer of this role (ie, at
baketime).

if we do want to do this later, gcenode has some yaml that is more
thorough in deleting files, we can add it back into this role under
control of a defaulted-false variable at that time.