+ callbacks: smemshy: copy of unixy from community.general c823e37d00
this copied straight from the git repo of the standard collection with
the unixy file current asof 20250116213501

did change unixy -> smemshy, but if this seems to work well over time,
should try to get it upstream.  that is, adding a few small features.

repo is at https://github.com/ansible-collections/community.general