+ unixy: do not display stdout or stderr if not actionable or verbose
if we are either ok or changed, don't display the stdout, stderr or msg
field.  this when not in verbose mode, which can be set with '-v' but
also some things like 'debug' module have this set.  depending on
display_ok_hosts, only 'name' will be printed and if changed, there is
also a 'done' line printed.  ideally this would be on the same line, but
second line only in changed case, is good enough for this pass

as long as every task has a good 'name' the output is concise will still
being useful when errors occur, and still displaying things when verbose
is set on play or task level