- roles: all: tasks: name: make sure this is set for unixy output
unixy (or really ansible, since unixy actually doesn't have the code to
know itself) substitutes the module name if the task has no name.  so
when we are only outputting for changed things, as we are doing with it,
this is not enough information.  all tasks should have names then, even
trivial ones, because otherwise there are just random module names
printed in the output with no way to know what was done with it, only
that the module ran