+ collections: add the incus collection from kmpm
this is basically the lxd one, but modified to work with incus.  this is
supposed to get into community.general eventually, for the meantime if
it works then whatever, we only use it for one thing: node shutdown
during image bake, and it's better than using 'command' to do it.  we
cannot naively run 'shutdown' on the host because it might not stop, so
we need the ability to force it.  terraform provider doesn't allow
shutdown either.