+ host: new role to place our flat hostdata, includes data files
these files (/etc/hosts, /etc/genders, /etc/resolv.conf
/etc/ssh/ssh_known_hosts) are not templated at this point, we have
static blocks of hosts, we know what they are, and everything is
configured statically.  there are some one-off pet snowflake servers and
they're recorded here, other than that we just use big ranges of
monotonically numbered host resources in a flat domain namespace.  we
can use the nodename itself to encode groupings.

this scales fairly well for a small setup, up to a certain level of
complexity, which should be fine for our needs at the moment. we use
10-bit network ranges so we have 1024k hosts in each range for any kind
of grouped type and we could have as many of those as we want.  for
small home setup this works fine, no need for dns or network management
tools.  static routes and build this as we go using terraform is a fine
strategy.

not sure it's good to maintain these primary data files with the role,
but at least it's via symlink into the crypt.  but in terms of
logically, the hostdata should probably be separate somewhere.  maybe
even moved to terraform and used here in the ansible via inventory with
terraform-provider-ansible? it's not clear to me that these are two way.
at the moment we define some things in two places, but only patterned
stuff such as omniplexN matching the N from the last octet in some of
these ranges (in groups of 256 at a time anyways... we can cover the /24
boundary but it would cease matching after the first octet)

anyways, we'll see how this works out for now and adjust strategy later
as needed.