^ lxdbake: upload_and_destroy: shutdown_bakenode: do not force stop
this is all under our control so we should not have to force it.  this
was done during debugging because there was some issue shutting it down,
but that actually had to do with terraform refreshing the state
prematurely, rather than some problem with the node itself.  so we'll
remove this for now and see if we need to use it again later.