/ lxdbake: avoid bakenode startup before imaging by not refreshing state
when terraform refreshes the state, even though we successfully shut it
down, it ends up being start up again as part of state check.  we want
to avoid this, so we pass "refresh false," which gives a
"-refresh=false" on the "plan" cli.  which at the moment requires my
patch in pull request 168 to the cloud.terraform.terraform module call.
but it solves the issue that the node can't be imaged when shut down,
and making an image starts it up by the status check.