+ collections: terraform: add the cloud.terraform collection
add submodule with cloud.terraform.terraform.  this is a fork of
the community.general.terraform collection.  the fork appears to be
maintained directly by redhat authors but isn't otherwise very distinct,
or have anything interesting in it, so no compelling reason to use.  we
have, however, made a change to it (adding -refresh= flag capability for
the module user), which is easier in a dedicated repo, so we have
switched to cloud.terraform, and pegged to by commit in my upstream PR
167.  it adds "refresh: false" possibility as a module arg.