- cloudinit: net: gateway4 is deprecated, use routes instead
shows deprecation warning in the logs, looked it up and this is how
we're supposed to do it now.  on later netplan (ie u24), it actually
accepts a string "default" but that won't be universal so we need to use
the actual catch-all address rather than a symbolic name that works on
only some netplan versions

(we have to use netplan, because that's the backend that parses v2
config.  however we will write the config using eni method in upcoming
commits, so we can use ifupdown proper and get rid of this newfangled
netplan stuff)