= cloudinit: user: header: make clear in comments this is for baketime
we're using an inline heredoc in terraform for the simpler case of an
already-baked image just having a new identity applied to it.  for that
use case we are not using this role.  so add a comment so that's
slightly more apparent than it was before (hopefully enough to trigger
my memory ;-)