+ cloudinit: user: remove netplan from activators and renderers
we want to use /etc/network/interfaces (eni method), so we just need to
(1) make sure ifupdown is installed, and (2) ensure that eni is earlier
in the list than other ones that would work on ubuntu (like netplan).
simplest way to do that is to just remove netplan from the relevant
cloud.cfg entry.

sadly, there seems to be no way to do this in a data source, it has to
be done in the config files.  this means there probably isn't a way to
do a one-stage bake, because the first stage will always put netplan
first in the priority list