/ tf: main: incus_storage_pool: set storage volume.size to match profile
even though the profile specifies 128 GiB, it doesn't seem to override
the volume size.  not sure why this is, but otherwise the image breaks
like this:

  Failed creating instance from image: Source image size (137438953472)
  exceeds specified volume size (10737418240)

this is 128GiB (the size when imaged) vs 10GiB (the configured volume
size).  we were thinking the 10GiB was a default that the profile could
override, but what it really seems to be is the initial volume size
itself.

since it has worked before with 10GiB but smaller images (this one is
only ~1G footprint, but expands to be a 128GiB filesystem), my guess is,
this is the initial size of the volume before the instance config has
even been considered, but then it gets booted with a growpart step to be
the size specified in the profile.

have to look into this more.  in the meantime we'll implement this fix
because it works and enables us to move on...