- host: interfaces: include cloud-init config rather than template ours
we are switching to ENI renderer in cloud-init.  this only writes to
/etc/network/interfaces.d/NN-cloud-init.cfg, which are actually not
all valid chars for interfaces(5) 'source-directory' directive (the dot
is not valid, so it doesn't get source).  so we have to use globbing and
include the specific file rather than all files in the include directory

first considered moving the cloud-init file into place (as
/etc/network/interfaces) after it gets rendered, but when is that? some
cursory searching didn't reveal if chicken or egg comes first, so for
the time being we'll just source the cloud-init config file and revisit
this later if it becomes bothersome.

at some point we might just have to consider using systemd-networkd
rather than continuing to make our lives difficult.  systemd has a lot
of authoritarian dain bramage but it's the reality for many years now
and not going anywhere.

or maybe we'll just switch to alpine and gentoo, or devuan.  we'll save
that for later...