/ tf: main: omniadmv: user-data: write /etc/hosts ourselves since no disable
cannot seem to figure out how to disable overwrite of /etc/hosts.  the
typical options in user-data are ignored.  tried many variations
including remove from cloud.cfg

however we do have a static one which is easy to obtain and write, so we
will just write_files it at this point and then everything is solved for
now.  no need to fix it until later if it comes up