/ callbacks: smemshy: further attempts to eliminate extra spaces
this whole thing is a hack with some spaces kept, some ignored, all so
the output has just one space.  it's not predictable... need to refactor
this and some of the functions it calls, bring them into the callback
locally so we can have our own version.  but right now there's no time,
we just want it to work, so we'll worry about it later