- tf: files: separate backend, locals, providers, variables, versions
reformatted into separate files to follow the style guide on hashicorp's
web site and also following the conventions described in several reddit
posts and stackexchange articles asking what is the proper technique.
this seems to be the most "contemporary" way.  we only started with the
single file to bootstrap the setup, it grew to about 400 lines so that's
a reasonable time to start separating it...