/ cfgnode: mkuser: give config user key in list to access config acct
not clear how this even worked before without providing the users list,
but obviously we used this role many many times and it worked fine.
it's breaking now, something must have changed but we just want the one
user for now, config user themselves to be able to access config
account.  when adding this the role works again.