- postfix: defaults: add variable for dnsroute cron interval
we want this to be variable... might change this later to 5m to reduce
mail latency in first-connect scenarios.  or maybe it won't cause a
problem in practice so we will leave it.  either way we have this
convenient variable now to use for if/when we choose to do that change