- ansible: use ansible-core only, with submodules for used collections
our plays right now use only 3 collections from the ansible pypi
package, yet it bundles a great many superfluous ones which takes a
while to install.  we're trying to get the venv rebuild time down, so,
after an audit, we found that this is probably easier, to just use
submodules and avoid installing ansible completely.

so to summarize, we remove ansible, install ansible-core only and add
submodules for the 3 collections we use