+ smemsh: repos: move pegs to defaults, and allow individual overrides
previously we had a static version peg in vars and would have always had
to edit the file if we wanted to change version numbers, it could not be
supplied at runtime.  we make some changes:

  - generalizes the logic so we don't have to have duplicate loop code
    for installrc vs installx, but can use the same loop, using inline
    jinja templating for variable construction in the yaml
  - uses the combine() filter to allow user to override specific
    versions using extra vars or other variable sources to provide
    skel_exes_addl or skel_rcfiles_addl, this way we can have defaults
    for usual run but allow to override specific ones at runtime on cli
    when the role is run
  - moves the static pegged versions out of vars/ and into defaults/
    since we can now override them

so, this is making the versions more dynamic.  in upcoming commits we
will make this fully dynamic by removing the pegs and allowing empty peg
values to get looked up automatically on github.  this patch lays some
groundwork for that.