+ smemsh: ghreltags: add filter+yaml to resolve git repos at latest tag
tired of changing the peg all the time and having to keep up any time
there is any little change.  we'll instead just check github using the
api.  have to include our key because the anonymous limit is very low,
but we vault it.  this role is likely to only ever used by me anyways so
may as well put my stuff in here...