+ smemsh: defaults: provide defaults so normal case is no-args
we had to use several extra args to get this role to work before,
because there was long ago an intention to make it so other people could
use it, but over time have come to realize nobody will use this except
me so might as well hardcode my own defaults.  after all, they could
still be overridden in extra args.