+ smemsh: gits: reduce api usage by doing github version checks just once
avoid talking to github on every host, to avoid using the api so much.
really these are always going to be the same so we'll just set the facts
locally for the localhost only, and refer to them explicitly by using
the copy via hostvars['localhost'], this way it's only done once per
play run rather than once per host.  this will be most useful when we
run an update and want to deploy it to all our hosts