/ smemsh: installx: osverchk: fix faulty conditional logic
this was a pretty dumb thinko, obviously the string was already
lowercase, we used parentheses on the wrong portion.  so this was never
actually being run.  it's not needed anymore anyways really since it's
basically certain we'd never be running ubuntu12 anymore ;-)
nonetheless, we'll leave this as a stub in case we need to do other
version checks later or support/test on non-ubuntu systems