+ smemsh: init: add skel_init var, default false, to run first-time roles
only on the first run do we really want to run all the initialization
roles.  even though it's idempotent and in theory we could run these all
every time, it's also a waste of time and we'll be doing this often
enough that it would be annoying.  so now we have to pass skel_init=true
to get it to run all these roles, otherwise it will skip right to the
git/installx stuff