^ rsyslog: rc: klog: remove no-op rule that falsely implies termination
after some research, discovered that this rule actually does nothing.
there is no consequence because all the rules have already been
traversed, but its intention was to drop any non-kernel message so it
wouldn't be further processed, however that is not actually what it
does: "kern.none ~" matches "all kernel facility messages, but also
"none of them with any priority level," which effectively matches
nothing.  the intention of "none" is to augment a previous "something"
such as: "*.;kern.none" which would mean "all non-kernel messages" and
act as an exclusion.

because this was misleading, and does nothing, it will be removed