- rsyslog: rc: klog: replace system processing of kern.log with ours
we previously used a separate file /var/log/klog (analogous to
/var/log/syslog) to write our own filtered log messages for klogpoll,
but that means it will never get rotated.  maybe that's a good thing,
but remember that we'll get important messages in the mail anyways, from
klogpoll.

instead, so we can get the existing rotation, we will write to kern.log,
which is already defined in 50-default.conf, and that file is already
rotated by the system (in /etc/logrotate.d/rsyslog). however, since our
klog processing is defined after 50-default.conf, in 90-klog.conf, we
need to change ours to be earlier, so in this patch the filename has
moved to 02-klog.conf.

furthermore, since 50-default will write all kernel messages to kern.log
itself, that would result in duplicates: we would do our own processing,
write some of the messages, and then 50-default would do it again.  we
don't want to modify the distro file, but we do want to handle
filtration of the kern.log file ourselves.  so what we now do is ALL
kern.* processing ourselves in our now-earlier-lexed config file.  We
have to write to /var/log/{syslog,messages,kern.log} ourselves, and we
do the latter two with filtration, the former without.

to avoid having to painstakingly define the actions for every rule, we
had to change syntax to make a block using conditional logic.

along the way some patterns got updated, for example to match the acutal
iptables prefixes that are extant on our systems.  essentially it's the
same though.  and we left some older rules commented out so we know how
to add, eg hostname condition, later on.  possibly we would do this by
templating the file in ansible though, and not inline, but we'll worry
about that later.