+ dmesg: new role that deposits klogpoll and sets up cron job
we'll use this on any system that runs a kernel (that is, not
containers).  it relies on mail working, cron existing, and python3.
deposits the klogpoll from utiladm, which we already have as a submodule
in this repo

for this role we need klogpoll to use the kern.log path we've been
working on in rsyslog role rather than our own separate file that we
used before.  this avoids having to maintain a logrotate script