^ vars: ipnet: removing this as no longer used and prefer inline
we're not using this anywhere in the repo except one reference in
vmiterate2 -- which is a wip -- and only to a single expansion variable
from this file.  while nice to have these, probably being explicit with
the ipaddr() module inline would be the better way to go in most cases.

needing to have these also potentially complicates some inventory
refactoring we're doing, in particular we're trying to get rid of
host_vars directory and replace with a single inventory file and each of
our plexhosts has an '.ipnet' member that we don't want.  while that
doesn't use this, different ways of calculating it would and we don't
want those to use explicit ipaddr() calls which removes possibly the
last reason to use these expansions.

we'll leave this in the archive directory for later reference or
revivial should the need arise...