^ hostdata: vars: plex: remove ipnet var, unused and undesired here
we don't want this to be statically defined via hardcodes in inventory.
especially all these silly files one per host.

the actual addresses are a bit of a challenge to acquire properly and we
want to do it from truth sources like the actual hostname, inline via a
lookup filter (which we have not yet written and doesn't exist already
in ansible, but is easy to write with socket.gethostbyaddr()) or
eventually maybe from terraform ansible provider resources.

possibly we will also use the hostname to figure out the plex group it
belongs to, which could get the plex net from networks.yml and we could
get it that way.

but we'll implement that later, for now best to remove these, so
host_vars can be factored out and we can switch to a yaml based
inventory