^ genders: remove vm/kos groupings from plexhocs as these can differ
any node can be either a system container or a virtual machine so we
won't hardcode these here except as one-offs when required.  possibly we
need these to be added based on composed inventory using something
coming from fact discovery.  but we'll leave that for the future if/when
it comes up