- genders: change 'vm' group to 'vme' in keeping with past practice
mainly so the plural isn't 'vms' which always reminds us of the old VAX
systems.  like also that it includes "environment" in the term, which
seems more descriptive.  anything can run in a VM, after all, but we
like to think of an "operating environment" which runs an "operating
system."  plus, we have used 'vme' for years in profiessional contexts
so this is muscle memory by now

words, words...