+ genders: reimplement inventory as native yaml, including host_vars
finally we can rid ourselves of these one-file-per-host host_vars which
becomes unwieldy at larger counts and involve needless duplication.
well, we could have used group_vars to factor out some duplication but
we still would have needed some source for the host entries themselves.

for a long time we have used the LLNL genders format as our primary
groups data, and imported from the static files with an inventory script
that would output JSON in ansible inventory format.

now we will do the opposite: use ansible native inventory, and generate
the genders file therefrom.  this is more flexible and allows us to use
composites of different inventory sources to create not only ansible
inventory but also the genders file for impromptu machine orchestration
by ssh.  importantly, we can also export terraform inventory into
ansible and that will end up in the genders file.

this commit only changes the inventory.  it's now 1:1 equivalent with
the crypt/hostfiles/genders (logically -- that is, after the command
"nodeattr --compress" is run on it) that was created by the bin/genders
script.

followup commits will reimplement the genders file as a template created
by inventory export, parse and compress.

we might also template the hosts file and parts of known_hosts while
we're at it.  this will mean less files to edit too... and sets the
stage for later on switching to DNSSEC+SSHFP to replace ssh_known_hosts