+ rc: inventory: plugins: pass extra vars to any plugin that can use them
this could be convenient when, for example using constructed inventory
sources and we need to use variables in vars_files, we can pass them on
the command line as "-e @/path/to/vars.yml" and still get them as we
would with an ansible play run.  this would rely on lazy evaluation
because normally those wouldn't be loaded until playtime, ie until after
inventory was parsed.  so it's not clear that that actually works, but
it might work, and in any case allowing the user to override with extra
vars is a good thing with little downside

note that only some plugins seem to support this, like the 'constructed'
one