+ bin: ansvarps: quick hack to display var precedence from documentation
constantly going to this page, just have a script to dump it in the
terminal to save me some time.  and use the online copy to make sure
it's fresh.  it'll spit nonsense or empty if the format changes so it
should be obvious when we'll need to update the script.