- ansrole: init: don't set inventory/config, just use rc, or cli override
setting inventory to /dev/null and forcing with '-i' supplied in the
wrapper was needed one time when we used to run ansible from a different
user than the home dir it was in.  however, we never do that anymore,
always using the setup user on the control node.  it's interfering with
genders file generation, because we run ansible-inventory and it's
getting ANSIBLE_INVENTORY=/dev/null.

this was never an issue before because ansrole was always using, for
example "-i <firstart>" internally from first of the colon-separated
inv:onto:role arg -- usually supplying 'etc', the inventory dir -- but
now that we run ansible-inventory to generate genders, then when that
happens under ansrole, we would also have to give a '-i' argument there,
or it would use the $ANSIBLE_INVENTORY environment, which is to say *no*
inventory, and generate an empty file (which it does! quite surprising).

instead we'll just rely on the actual inventory configuration and stop
passing it.  we can always override with '-i' instead of always having
it supplied within the wrapper script.