/ ansrole: usagex: document the -n/--nofacts option in usage message
this was documented in the header, but not the actual usage output.
it's too bad we have to maintain this usage in two places.  we should
fix that by moving to the devskel template, which can handle this better
by displaying a portion of the header as the usage message