+ ansrole: change to a 2-arg invocation, using rc-configured inventory
instead of <inventory>:<patterns>:<role> as the primary argument
specifier (outside flags, ansible-playbook pass-throughs and variables
to pass in extra args), instead we can simplify things, and allow more
flexibility, by using a <patterns> <role> syntax, and using the stock
ansible-configured inventory by default.  if the user wants, they can
use --inventory to override, like they would with ansible-playbook by
itself.  this has the advantage of having unambiguous <pattern> string
because it will exactly match the documentation for targeting pattern
strings.  it also allows the implementation to be considerably simpler
because we don't have to do any regex matching, just pass the paramter
along unchanged.  and we can specify whatever --inventory args we want,
to get whatever subset of inventory we want to target, while honoring
the typical case that we just want to use our "usual" inventory and not
have to specify 'etc' all the time (ie "etc:omnius:ping" is now just
"omnius ping").

the regression is that we used to give the option of a simpler
<inventory>:<role> syntax and auto-convert it into a host_list-plugin
conformant inventory, allowing the user to target arbitrary hostnames
that might not have been in inventory.  however this only saved a
trailing comma and a need to give '--inventory', but it really didn't
save much compared to how often it's used.  it becomes:

  old: "host1:role"
  new: "--inventory=host1, all role"

that's not very good or intuitive, but that's in the unusual case where
we don't want to use our own inventory and just want to target some
random host (probably one that isn't ours, like for running smemsh role
to setup our user).

so, there's some problems still: for one, that we have to use
"--inventory=host1," and can't use "--inventory host1" or "-i host1"
(due to the problem with spaces) but this is a long-standing issue with
our script not doing proper argument parsing, and isn't really a
regression per se.

for two, now we have to use a trailing comma to activate host_list
processing.  however, this is the same as how ansible-playbook would be
itself anyways.  it does, however, require the user to "know" that.
probably the best way around this would be to provide our own special
argument (like we use '-n/--nofacts') when we want to target just raw
hostnames and rewrite it ourselves into the ansible format.

but we'll consider these remaining issues in later commits...