+ ansrole: switch to devskel/getopt for arg processing and general structure
this gives us a way to have our own arguments out of order and take
arguments if we want.  they still have to be at the beginning:
ansible-playbook options still need to trail the inventory spec.  but it
removes the need to use exclusively the --longopt=value options, we can
just use "-i foo" now, we could not do it before because we only
processed args with a leading '-'.  and also we'll get the benefit of
tracking devskel changes upstream and potentially bringing them in.