+ ansrole: args: wrap playbook option processing, allows intermix/rename
this will enable us to:

  - have our own name for options (usually, shorter longopt names)
  - our own option processing conventions, eg to change option semantics
  - intersperse our own script options with ansible-playbook options
  - allow help to be written how we like, format and options printed

we could not do this otherwise, because getopt(1) stops on the first
non-option argument encountered.  we have to process later var=value
options (after the pattern and role args), so even if we required the
user to put ansible-playbook options afterwards, it would still have
edge cases, for example if the user put "-e var=value" as the last one
before other "var=value" for the script to parse natively, it would have
to be taught that '-e' takes an argument, and possibly other edge cases,
and then we're halfway to option processing anyways.

we had first considered parsing "ansible-playbook --help" but that has
some differences with the actual options (such as '-b' not shown in the
summary line, only '--become'), and also takes about 0.33 seconds to
run on its own, which we don't want to incur each and every time
(because we still have to incur it a second time when we run the actual
ansible-playbook line).

even despite these caveats, and having to maintain a hardcoded list in
triplicate (comment header for usage, getopt line, and the main arg
processing switch), we like the idea of supporting only a subset of the
options and possibly renaming them and giving our own semantics, it
opens up a lot of possibilities.  so this seems like a better way to go
for the moment.  a bit arduous, and we'll have to stay on top of any new
options ansible has, but we wouldn't know about those until discovering
them anyways and we can always add them later now that we have a
convention for representing them in the script.

we leave only the common ones that we might use and/or have used in the
past.  if the need for others should arise we'll add them at that time.