/ ansrole: args: if user specifies inventory, do not skip pdsh_all_skip
if the user is giving an inventory argument, we will let them have full
control and not make the assumption that a pdsh_all_skip group exists.
chances are this is for specifying a host_list instead of the default
inventory, and that would not have such a group configured.  in general
that hardcoded assumption should go away... we should probably have an
'all' group in ineventory itself that isn't really 'all' but actually
the value of $allnodes already, and we could use the "compose" inventory
plugin for this, to make a group called "vall" maybe that was just the
same as "all:!pdsh_all_skip" but we will leave this for later and think
about if that's the best way to implement it